feat(#42):solved change requests
