+++ src/app/api/user/route.ts
@@ -60,29 +60,4 @@ export async function POST(request: NextRequest) {
         message: "Nutzer erfolgreich angelegt"
     })
 }
-export async function DELETE(request: NextRequest) {
-    const supabaseClient = createAdminClient();
-    const data = (await request.json()).formData as User;
-    const isTrader = await IsTrader(supabaseClient);
-    if (!isTrader) {
-        return NextResponse.json({ message: "Nicht autorisiert" }, {
-            status: 401,
-        });
-    }
 
-    const { data: { user }, error } = await supabaseClient
-        .auth
-        .admin
-        .deleteUser(data.id);
-
-    if (error) {
-        return NextResponse.json({ message: "Fehler beim Löschen des Nutzers " + error.message }, {
-            status: 500,
-        });
-    }
-
-    return NextResponse.json({
-        data: user,
-        message: "Nutzer erfolgreich gelöscht"
-    })
-}
